Clean up SearchQuery API (less misleading names, fluent interface)
SearchQuery->search() is highly misleading - it implies that a search is executed,
while in fact it only adds a search term to an internal property.
This gets even more confusing since the actual method executing passing
on the search request to the backend is also called SearchIndex->search().

Renamed other methods along the same lines, for example start() is a confusing
name for adding a start value for pagination, and inClass() implies a boolean
return while in fact its adding a filter.

Also added getters to get naming a bit more consistent (e.g. filter() sets $require).

Set the deprecation notices to 2.0, so a future release.