Remove commit as an abstract feature. Rely on auto-commits on Solr.
Commit is not really a backend-independent feature, so remove it from
abstract classes such as SearchIndex.

Also, Solr4 preferred way to deal with commits is server-side
auto-commit. Client-driven commits are risky because they can cause too
many commits running in parallel, which can cause performance issues
on the backend (such as the maxWarmingSearchers error).

Note the commit capability is retained in the Solr4Service, in case
a power-user with good undestanding of solr internals needed a more
fine-grained control over commits.