FIX permit reindexing on a Windows machine
It appears that when executing in a Microsoft Windows environment the
default shell used in the immediate reindex handler is CMD, which has
some severely different syntax support from that of a POSIX machine (e.g.
GNU bash). This leads to errors to do with double escaping certain
parameters passed in via the CLI - culminating in an Injector error about
an non-extant class e.g. 'SilverStripe\\CMS\\Model\\SiteTree' - two
slashes a namespace separator does not make!